Set up the code path to fetch recipe list

+++ app/src/main/java/info/eric/nobaking/model/Ingredient.java
@@ -7,7 +7,7 @@ import com.squareup.moshi.Moshi;
 @AutoValue
 public abstract class Ingredient {
 
-  public abstract int quantity();
+  public abstract float quantity();
 
   public abstract String measure();
